# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Forge Core API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memory-items:
    get:
      operationId: listMemoryItems
      tags: [memory]
      summary: List / search memory items
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string }
        - name: category
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Memory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryItem"
    post:
      operationId: createMemoryItem
      tags: [memory]
      summary: Create memory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryItem"
  /memory-items/{id}:
    delete:
      operationId: deleteMemoryItem
      tags: [memory]
      summary: Delete memory item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /improvements:
    get:
      operationId: listImprovements
      tags: [improvements]
      summary: List self-improvement backlog
      responses:
        "200":
          description: Improvements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Improvement"
    post:
      operationId: createImprovement
      tags: [improvements]
      summary: Create improvement item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImprovementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Improvement"
  /improvements/{id}:
    patch:
      operationId: updateImprovement
      tags: [improvements]
      summary: Update improvement (e.g. priority)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImprovementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Improvement"
  /improvements/{id}/convert:
    post:
      operationId: convertImprovementToTask
      tags: [improvements]
      summary: Convert improvement into a module task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Task created from improvement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  schemas:
...
    Task:
      type: object
      required: [id, title, status, risk, ownerAgent, createdAt]
      properties:
        id: { type: integer }
        projectId: { type: ["integer", "null"] }
        parentTaskId: { type: ["integer", "null"] }
        title: { type: string }
        goal: { type: ["string", "null"] }
        scope: { type: ["string", "null"] }
        risk: { type: string }
        ownerAgent: { type: string }
        status:
          type: string
          enum: [draft, planned, active, blocked, testing, review, approved, rejected, done]
        acceptanceCriteria: { type: ["string", "null"] }
        blockedReason: { type: ["string", "null"] }
        source: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: ["string", "null"] }
...
    MemoryItem:
      type: object
      required: [id, category, title, content, createdAt]
      properties:
        id: { type: integer }
        category:
          type: string
          enum:
            - successful_module
            - failed_module
            - error
            - solution
            - architecture_choice
            - dependency_issue
            - test_result
            - recurring_blockade
            - lesson_learned
        title: { type: string }
        content: { type: string }
        tags:
          type: array
          items: { type: string }
        taskId: { type: ["integer", "null"] }
        createdAt: { type: string }
    MemoryItemInput:
      type: object
      required: [category, title, content]
      properties:
        category:
          type: string
          enum:
            - successful_module
            - failed_module
            - error
            - solution
            - architecture_choice
            - dependency_issue
            - test_result
            - recurring_blockade
            - lesson_learned
        title: { type: string, minLength: 1 }
        content: { type: string, minLength: 1 }
        tags:
          type: array
          items: { type: string }
        taskId: { type: integer }
    Improvement:
      type: object
      required: [id, problem, priority, status, createdAt]
      properties:
        id: { type: integer }
        problem: { type: string }
        cause: { type: ["string", "null"] }
        proposedModule: { type: ["string", "null"] }
        expectedImprovement: { type: ["string", "null"] }
        risk: { type: string }
        priority: { type: string }
        requiredTests: { type: ["string", "null"] }
        status: { type: string }
        source: { type: ["string", "null"] }
        createdAt: { type: string }
    ImprovementInput:
      type: object
      required: [problem]
      properties:
        problem: { type: string, minLength: 1 }
        cause: { type: string }
        proposedModule: { type: string }
        expectedImprovement: { type: string }
        risk: { type: string }
        priority: { type: string }
        requiredTests: { type: string }
        source: { type: string }
    ImprovementUpdate:
      type: object
      properties:
        problem: { type: string, minLength: 1 }
        cause: { type: string }
        proposedModule: { type: string }
        expectedImprovement: { type: string }
        risk: { type: string }
        priority: { type: string }
        requiredTests: { type: string }
        status: { type: string }
...